made changes to the player client to reflect unity of game server / superpeer
